Fix team members server route boundary
